Settings panel

+++ sources/components/preferences/src/main/kotlin/com/sedsoftware/blinktracker/components/preferences/store/BlinkPreferencesStore.kt
@@ -8,7 +8,7 @@ import com.sedsoftware.blinktracker.components.preferences.store.BlinkPreference
 internal interface BlinkPreferencesStore : Store<Intent, State, Label> {
 
     sealed class Intent {
-        data class MinimalThresholdChanged(val value: Int) : Intent()
+        data class MinimalThresholdChanged(val value: Float) : Intent()
         data class NotifySoundChanged(val value: Boolean) : Intent()
         data class NotifyVibrationChanged(val value: Boolean) : Intent()
         data class LaunchMinimizedChanged(val value: Boolean) : Intent()
@@ -17,7 +17,7 @@ internal interface BlinkPreferencesStore : Store<Intent, State, Label> {
     }
 
     data class State(
-        val minimalMinuteThreshold: Int = -1,
+        val minimalMinuteThreshold: Float = -1f,
         val notifySound: Boolean = false,
         val notifyVibration: Boolean = false,
         val launchMinimized: Boolean = false,
